[refactor] game features are refactored
