+++ src/molecule/GameOnMatching.tsx
@@ -3,6 +3,7 @@ import {useEffect} from "react";
 import {Socket} from "socket.io-client";
 
 import Button from "@/atom/Button";
+
 import {currentGameStatus} from "@/states/game/currentGameStatus";
 import {matchInfo} from "@/states/game/matchInfo";
 
@@ -54,14 +55,11 @@ const GameOnMatching = (props: GameOnMatchingProps) => {
     return () => {
       socket.removeAllListeners('player_matched');
       socket.removeAllListeners('user_joined_room');
+      socket.removeAllListeners('user_exit_room');
       console.log("GameOnMatching unmounted");
     }
 
   }, []);
-
-  const gameMatched = () => {
-    setGameStatus("MATCHED");
-  }
 
   const cancelMatching = () => {
     socket.emit("user_left_queue");
@@ -71,11 +69,12 @@ const GameOnMatching = (props: GameOnMatchingProps) => {
 
   return (
     <div className="flex flex-col items-center justify-center h-full w-full stop-dragging">
-      <span className="text-white font-pixel text-2xl"
-      onClick={gameMatched}> Matching... </span> {/* TODO: 서버에 붙으면 OnClick 이벤트 삭제 */}
+      <span className="text-white font-pixel text-2xl">
+        Matching...
+      </span>
       <Button onClick={cancelMatching}
               className="bg-red-500 text-white font-pixel mt-10">
-                cancel
+        cancel
       </Button>
     </div>
   )
