+++ src/states/game/currentGameStatus.ts
@@ -1,5 +1,5 @@
 import {atom} from "recoil";
-import {GameStatus} from "@/pages/GameContainer";
+import {GameStatus} from "@/templates/GameContainer";
 
 export const currentGameStatus = atom<GameStatus>({
   key: "currentGameStatus",
